+++ projects/js-packages/licensing/components/activation-screen-controls/index.jsx
@@ -3,8 +3,8 @@
  */
 import { Button, TextControl } from '@wordpress/components';
 import { createInterpolateElement } from '@wordpress/element';
-import { JetpackLogo } from '@automattic/jetpack-components';
-import { sprintf, __ } from '@wordpress/i18n';
+import { JetpackLogo, Spinner } from '@automattic/jetpack-components';
+import { __ } from '@wordpress/i18n';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { Icon, warning } from '@wordpress/icons';
@@ -19,7 +19,7 @@ import './style.scss';
  *
  * @param {object} props -- The properties.
  * @param {Function} props.activateLicense -- function to handle submitting a license
- * @param {boolean} props.disabled -- should the controls be disabled
+ * @param {boolean} props.isActivating -- should the controls be disabled
  * @param {string} props.license -- the license code to edit or submit
  * @param {?string} props.licenseError -- any error that occurred while activating a license
  * @param {Function} props.onLicenseChange -- function to handle changes to license
@@ -27,7 +27,7 @@ import './style.scss';
  * @returns {React.Component} The `ActivationScreenControls` component.
  */
 const ActivationScreenControls = props => {
-	const { activateLicense, disabled, license, licenseError, onLicenseChange, siteUrl } = props;
+	const { activateLicense, isActivating, license, licenseError, onLicenseChange } = props;
 
 	const hasLicenseError = licenseError !== null && licenseError !== undefined;
 
@@ -35,16 +35,12 @@ const ActivationScreenControls = props => {
 		<div className="jp-license-activation-screen-controls">
 			<div className="jp-license-activation-screen-controls--content">
 				<JetpackLogo showText={ false } height={ 48 } />
-				<h1>{ __( 'Activate a product', 'jetpack' ) }</h1>
+				<h1>{ __( 'Add a license key', 'jetpack' ) }</h1>
 				<p>
 					{ createInterpolateElement(
-						sprintf(
-							/* translators: "%s" is the url of the site i.e. hopeful-weevil.jurassic.ninja. */
-							__(
-								'Enter the license key we sent to your email to activate your product for <strong>%s</strong>',
-								'jetpack'
-							),
-							siteUrl
+						__(
+							'<strong>Check your email</strong> for your license key. You should have received it after making your purchase.',
+							'jetpack'
 						),
 						{
 							strong: <strong></strong>,
@@ -58,10 +54,9 @@ const ActivationScreenControls = props => {
 							: 'jp-license-activation-screen-controls--license-field-with-error'
 					}
 					label={ __( 'License key', 'jetpack' ) }
-					placeholder="jp-Product34623432423423"
 					value={ license }
 					onChange={ onLicenseChange }
-					disabled={ disabled }
+					disabled={ isActivating }
 				/>
 				{ hasLicenseError && (
 					<div className="jp-license-activation-screen-controls--license-field-error">
@@ -74,9 +69,8 @@ const ActivationScreenControls = props => {
 				<Button
 					className="jp-license-activation-screen-controls--button"
 					onClick={ activateLicense }
-					disabled={ license.length <= 0 || disabled }
 				>
-					{ __( 'Activate', 'jetpack' ) }
+					{ isActivating ? <Spinner /> : __( 'Activate', 'jetpack' ) }
 				</Button>
 			</div>
 		</div>
@@ -85,7 +79,7 @@ const ActivationScreenControls = props => {
 
 ActivationScreenControls.propTypes = {
 	activateLicense: PropTypes.func.isRequired,
-	disabled: PropTypes.bool.isRequired,
+	isActivating: PropTypes.bool.isRequired,
 	license: PropTypes.string.isRequired,
 	licenseError: PropTypes.string,
 	onLicenseChange: PropTypes.func.isRequired,