+++ projects/js-packages/licensing/components/activation-screen-controls/test/component.jsx
@@ -30,30 +30,6 @@ describe( 'ActivationScreenControls', () => {
 		} );
 	} );
 
-	describe( 'Render the ActivationScreenControls disabled', () => {
-		const testProps = {
-			activateLicense: () => null,
-			disabled: true,
-			license: 'test',
-			onLicenseChange: () => null,
-			siteUrl: 'jetpack.com',
-		};
-
-		const wrapper = shallow( <ActivationScreenControls { ...testProps } /> );
-
-		it( 'controls are correctly disabled', () => {
-			const licenseField = wrapper.find( '.jp-license-activation-screen-controls--license-field' );
-
-			expect( licenseField ).to.have.lengthOf( 1 );
-			expect( licenseField.prop( 'disabled' ) ).to.equal( true );
-
-			const activateButton = wrapper.find( '.jp-license-activation-screen-controls--button' );
-
-			expect( activateButton ).to.have.lengthOf( 1 );
-			expect( activateButton.prop( 'disabled' ) ).to.equal( true );
-		} );
-	} );
-
 	describe( 'Render the ActivationScreenControls with license error', () => {
 		const testProps = {
 			activateLicense: () => null,