+++ projects/js-packages/webpack-config/src/webpack.js
@@ -1,3 +1,4 @@
+const fs = require( 'fs' );
 const path = require( 'path' );
 const webpack = require( 'webpack' );
 
@@ -8,6 +9,8 @@ const DuplicatePackageCheckerWebpackPlugin = require( 'duplicate-package-checker
 const MiniCssExtractPlugin = require( 'mini-css-extract-plugin' );
 const MiniCSSWithRTLPlugin = require( './webpack/mini-css-with-rtl' );
 const WebpackRTLPlugin = require( '@automattic/webpack-rtl-plugin' );
+const I18nLoaderWebpackPlugin = require( '@automattic/i18n-loader-webpack-plugin' );
+const I18nCheckWebpackPlugin = require( '@automattic/i18n-check-webpack-plugin' );
 
 const MyCssMinimizerPlugin = options => new CssMinimizerPlugin( options );
 
@@ -18,13 +21,14 @@ const isDevelopment = ! isProduction;
 const mode = isProduction ? 'production' : 'development';
 const devtool = isProduction ? false : 'eval-cheap-module-source-map';
 const output = {
-	filename: '[name].min.js',
-	chunkFilename: '[name]-[id].H[contenthash:20].min.js',
+	filename: '[name].js',
+	chunkFilename: '[name].js?minify=false&ver=[contenthash]',
 };
 const optimization = {
 	minimize: isProduction,
 	minimizer: [ TerserPlugin(), MyCssMinimizerPlugin() ],
 	concatenateModules: false,
+	emitOnErrors: true,
 };
 const resolve = {
 	extensions: [ '.js', '.jsx', '.ts', '.tsx', '...' ],
@@ -47,7 +51,13 @@ const MomentLocaleIgnorePlugin = () => [
 	} ),
 ];
 
-const MyMiniCssExtractPlugin = options => [ new MiniCssExtractPlugin( options ) ];
+const MyMiniCssExtractPlugin = options => [
+	new MiniCssExtractPlugin( {
+		filename: '[name].css',
+		chunkFilename: '[name].css?minify=false&ver=[contenthash]',
+		...options,
+	} ),
+];
 
 const MyMiniCssWithRtlPlugin = options => [ new MiniCSSWithRTLPlugin( options ) ];
 
@@ -59,7 +69,49 @@ const DuplicatePackageCheckerPlugin = options => [
 
 const DependencyExtractionPlugin = options => [ new DependencyExtractionWebpackPlugin( options ) ];
 
+const I18nLoaderPlugin = options => [ new I18nLoaderWebpackPlugin( options ) ];
+
+const i18nFilterFunction = file => {
+	if ( ! /\.(?:jsx?|tsx?|cjs|mjs|svelte)$/.test( file ) ) {
+		return false;
+	}
+	const i = file.lastIndexOf( '/node_modules/' ) + 14;
+	return i < 14 || file.startsWith( '@automattic/', i );
+};
+const I18nCheckPlugin = options => {
+	const opts = { filter: i18nFilterFunction, ...options };
+	if ( typeof opts.expectDomain === 'undefined' ) {
+		let dir = process.cwd(),
+			olddir;
+		do {
+			const file = path.join( dir, 'composer.json' );
+			if ( fs.existsSync( file ) ) {
+				const cfg = JSON.parse( fs.readFileSync( file, { encoding: 'utf8' } ) );
+				if ( cfg.extra ) {
+					if ( cfg.extra.textdomain ) {
+						opts.expectDomain = cfg.extra.textdomain;
+					} else if ( cfg.extra[ 'wp-plugin-slug' ] ) {
+						opts.expectDomain = cfg.extra[ 'wp-plugin-slug' ];
+					} else if ( cfg.extra[ 'wp-theme-slug' ] ) {
+						opts.expectDomain = cfg.extra[ 'wp-theme-slug' ];
+					}
+				}
+				break;
+			}
+
+			olddir = dir;
+			dir = path.dirname( dir );
+		} while ( dir !== olddir );
+	}
+	return [ new I18nCheckWebpackPlugin( opts ) ];
+};
+I18nCheckPlugin.defaultFilter = i18nFilterFunction;
+
 const StandardPlugins = ( options = {} ) => {
+	if ( typeof options.I18nCheckPlugin === 'undefined' && isDevelopment ) {
+		options.I18nCheckPlugin = false;
+	}
+
 	return [
 		...( options.DefinePlugin === false ? [] : DefinePlugin( options.DefinePlugin ) ),
 		...( options.MomentLocaleIgnorePlugin === false
@@ -78,32 +130,16 @@ const StandardPlugins = ( options = {} ) => {
 		...( options.DependencyExtractionPlugin === false
 			? []
 			: DependencyExtractionPlugin( options.DependencyExtractionPlugin ) ),
+		...( options.I18nLoaderPlugin === false ? [] : I18nLoaderPlugin( options.I18nLoaderPlugin ) ),
+		...( options.I18nCheckPlugin === false ? [] : I18nCheckPlugin( options.I18nCheckPlugin ) ),
 	];
 };
 
-/****** Module rules and loaders ******/
+/****** Module rules ******/
 
 const TranspileRule = require( './webpack/transpile-rule' );
+const CssRule = require( './webpack/css-rule' );
 const FileRule = require( './webpack/file-rule' );
-const MiniCssExtractLoader = options => ( {
-	loader: MiniCssExtractPlugin.loader,
-	options: options,
-} );
-const CssLoader = options => ( {
-	loader: require.resolve( 'css-loader' ),
-	options: {
-		// By default we do not want css-loader to try to handle absolute paths.
-		url: { filter: urlpath => ! urlpath.startsWith( '/' ) },
-		...options,
-	},
-} );
-const CssCacheLoader = options => ( {
-	loader: require.resolve( 'cache-loader' ),
-	options: {
-		cacheDirectory: path.resolve( '.cache/css-loader' ),
-		...options,
-	},
-} );
 
 // Note: For this cjs module to be used with named exports in an mjs context, modules.exports
 // needs to contain only simple variables like `a` or `a: b`. Define anything more complex
@@ -129,10 +165,9 @@ module.exports = {
 	WebpackRtlPlugin: MyWebpackRtlPlugin,
 	DependencyExtractionPlugin,
 	DuplicatePackageCheckerPlugin,
+	I18nLoaderPlugin,
 	// Module rules and loaders.
 	TranspileRule,
+	CssRule,
 	FileRule,
-	MiniCssExtractLoader,
-	CssLoader,
-	CssCacheLoader,
 };