+++ projects/packages/assets/src/class-assets.php
@@ -7,7 +7,9 @@
 
 namespace Automattic\Jetpack;
 
+use Automattic\Jetpack\Assets\Semver;
 use Automattic\Jetpack\Constants as Jetpack_Constants;
+use InvalidArgumentException;
 
 /**
  * Class Assets
@@ -19,6 +21,7 @@ class Assets {
 	 * @var array
 	 */
 	private $defer_script_handles = array();
+
 	/**
 	 * The singleton instance of this class.
 	 *
@@ -27,11 +30,21 @@ class Assets {
 	protected static $instance;
 
 	/**
+	 * The registered textdomain mappings.
+	 *
+	 * @var array `array( mapped_domain => array( string target_domain, string target_type, string semver ) )`.
+	 */
+	private static $domain_map = array();
+
+	/**
 	 * Constructor.
 	 *
 	 * Static-only class, so nothing here.
 	 */
 	private function __construct() {}
+
+	// ////////////////////
+	// region Async script loading
 
 	/**
 	 * Get the singleton instance of the class.
@@ -86,6 +99,27 @@ class Assets {
 	}
 
 	/**
+	 * A helper function that lets you enqueue scripts in an async fashion.
+	 *
+	 * @param string $handle        Name of the script. Should be unique.
+	 * @param string $min_path      Minimized script path.
+	 * @param string $non_min_path  Full Script path.
+	 * @param array  $deps           Array of script dependencies.
+	 * @param bool   $ver             The script version.
+	 * @param bool   $in_footer       Should the script be included in the footer.
+	 */
+	public static function enqueue_async_script( $handle, $min_path, $non_min_path, $deps = array(), $ver = false, $in_footer = true ) {
+		$assets_instance = self::instance();
+		$assets_instance->add_async_script( $handle );
+		wp_enqueue_script( $handle, self::get_file_url_for_environment( $min_path, $non_min_path ), $deps, $ver, $in_footer );
+	}
+
+	// endregion .
+
+	// ////////////////////
+	// region Utils
+
+	/**
 	 * Given a minified path, and a non-minified path, will return
 	 * a minified or non-minified file URL based on whether SCRIPT_DEBUG is set and truthy.
 	 *
@@ -135,22 +169,6 @@ class Assets {
 		 * @param string $non_min_path The non-minified path.
 		 */
 		return apply_filters( 'jetpack_get_file_for_environment', $url, $min_path, $non_min_path );
-	}
-
-	/**
-	 * A helper function that lets you enqueue scripts in an async fashion.
-	 *
-	 * @param string $handle        Name of the script. Should be unique.
-	 * @param string $min_path      Minimized script path.
-	 * @param string $non_min_path  Full Script path.
-	 * @param array  $deps           Array of script dependencies.
-	 * @param bool   $ver             The script version.
-	 * @param bool   $in_footer       Should the script be included in the footer.
-	 */
-	public static function enqueue_async_script( $handle, $min_path, $non_min_path, $deps = array(), $ver = false, $in_footer = true ) {
-		$assets_instance = self::instance();
-		$assets_instance->add_async_script( $handle );
-		wp_enqueue_script( $handle, self::get_file_url_for_environment( $min_path, $non_min_path ), $deps, $ver, $in_footer );
 	}
 
 	/**
@@ -270,6 +288,11 @@ class Assets {
 		return $ret;
 	}
 
+	// endregion .
+
+	// ////////////////////
+	// region Webpack-built script registration
+
 	/**
 	 * Register a Webpack-built script.
 	 *
@@ -286,17 +309,18 @@ class Assets {
 	 * @param string $path        Minimized script path.
 	 * @param string $relative_to File that `$path` is relative to. Pass `__FILE__`.
 	 * @param array  $options     Additional options:
-	 *  - `nonmin_path`:      (string) Non-minified script path.
 	 *  - `asset_path`:       (string|null) `.asset.php` to load. Default is to base it on `$path`.
+	 *  - `async`:            (bool) Set true to register the script as async, like `Assets::enqueue_async_script()`
+	 *  - `css_dependencies`: (string[]) Additional style dependencies to queue.
 	 *  - `css_path`:         (string|null) `.css` to load. Default is to base it on `$path`.
 	 *  - `dependencies`:     (string[]) Additional script dependencies to queue.
-	 *  - `css_dependencies`: (string[]) Additional style dependencies to queue.
-	 *  - `version`:          (string) Override the version from the `asset_path` file.
-	 *  - `minify`:           (bool|null) Set true to pass `minify=true` in the query string, or `null` to suppress the normal `minify=false`.
 	 *  - `enqueue`:          (bool) Set true to enqueue the script immediately.
-	 *  - `async`:            (bool) Set true to register the script as async, like `Assets::enqueue_async_script()`
 	 *  - `in_footer`:        (bool) Set true to register script for the footer.
 	 *  - `media`:            (string) Media for the css file. Default 'all'.
+	 *  - `minify`:           (bool|null) Set true to pass `minify=true` in the query string, or `null` to suppress the normal `minify=false`.
+	 *  - `nonmin_path`:      (string) Non-minified script path.
+	 *  - `textdomain`:       (string) Text domain for the script. Required if the script depends on wp-i18n.
+	 *  - `version`:          (string) Override the version from the `asset_path` file.
 	 * @throws \InvalidArgumentException If arguments are invalid.
 	 */
 	public static function register_script( $handle, $path, $relative_to, array $options = array() ) {
@@ -308,14 +332,15 @@ class Assets {
 		$base     = substr( $path, 0, -3 );
 		$options += array(
 			'asset_path'       => "$base.asset.php",
-			'css_path'         => "$base.css",
-			'css_dependencies' => array(),
-			'dependencies'     => array(),
-			'minify'           => false,
-			'enqueue'          => false,
 			'async'            => false,
+			'css_dependencies' => array(),
+			'css_path'         => "$base.css",
+			'dependencies'     => array(),
+			'enqueue'          => false,
 			'in_footer'        => false,
 			'media'            => 'all',
+			'minify'           => false,
+			'textdomain'       => null,
 		);
 
 		if ( $options['css_path'] && substr( $options['css_path'], -4 ) !== '.css' ) {
@@ -352,6 +377,16 @@ class Assets {
 		wp_register_script( $handle, $url, $options['dependencies'], $ver, $options['in_footer'] );
 		if ( $options['async'] ) {
 			self::instance()->add_async_script( $handle );
+		}
+		if ( $options['textdomain'] ) {
+			wp_set_script_translations( $handle, $options['textdomain'] );
+		} elseif ( in_array( 'wp-i18n', $options['dependencies'], true ) ) {
+			_doing_it_wrong(
+				__METHOD__,
+				/* translators: %s is the script handle. */
+				esc_html( sprintf( __( 'Script "%s" depends on wp-i18n but does not specify "textdomain"', 'jetpack' ), $handle ) ),
+				''
+			);
 		}
 
 		if ( $options['css_path'] && file_exists( "$dir/{$options['css_path']}" ) ) {
@@ -391,4 +426,274 @@ class Assets {
 		}
 	}
 
+	/**
+	 * 'wp_default_scripts' action handler.
+	 *
+	 * This registers the `wp-jp-i18n-state` script for use by Webpack bundles built with
+	 * `@automattic/i18n-loader-webpack-plugin`.
+	 *
+	 * @since 1.14.0
+	 * @param \WP_Scripts $wp_scripts WP_Scripts instance.
+	 */
+	public static function wp_default_scripts_hook( $wp_scripts ) {
+		$data = array(
+			'baseUrl'   => false,
+			'locale'    => determine_locale(),
+			'domainMap' => array(),
+		);
+
+		$lang_dir = Jetpack_Constants::get_constant( 'WP_LANG_DIR' );
+		$abspath  = Jetpack_Constants::get_constant( 'ABSPATH' );
+
+		if ( strpos( $lang_dir, $abspath ) === 0 ) {
+			$data['baseUrl'] = site_url( substr( trailingslashit( $lang_dir ), strlen( untrailingslashit( $abspath ) ) ) );
+		}
+
+		foreach ( self::$domain_map as $from => list( $to, $type ) ) {
+			$data['domainMap'][ $from ] = ( 'core' === $type ? '' : "{$type}/" ) . $to;
+		}
+
+		/**
+		 * Filters the i18n state data for use by Webpack bundles built with
+		 * `@automattic/i18n-loader-webpack-plugin`.
+		 *
+		 * @since 1.14.0
+		 * @package assets
+		 * @param array $data The state data to generate. Expected fields are:
+		 *  - `baseUrl`: (string|false) The URL to the languages directory. False if no URL could be determined.
+		 *  - `locale`: (string) The locale for the page.
+		 *  - `domainMap`: (string[]) A mapping from Composer package textdomains to the corresponding
+		 *    `plugins/textdomain` or `themes/textdomain` (or core `textdomain`, but that's unlikely).
+		 */
+		$data = apply_filters( 'jetpack_i18n_state', $data );
+
+		if ( ! is_array( $data ) ||
+			! isset( $data['baseUrl'] ) || ! ( is_string( $data['baseUrl'] ) || false === $data['baseUrl'] ) ||
+			! isset( $data['locale'] ) || ! is_string( $data['locale'] ) ||
+			! isset( $data['domainMap'] ) || ! is_array( $data['domainMap'] )
+		) {
+			$js = 'console.warn( "I18n state deleted by jetpack_i18n_state hook" );';
+		} elseif ( ! $data['baseUrl'] ) {
+			$js = 'console.warn( "Failed to determine languages base URL. Is WP_LANG_DIR in the WordPress root?" );';
+		} else {
+			$data['domainMap'] = (object) $data['domainMap']; // Ensure it becomes a json object.
+			$js                = 'wp.jpI18nState = ' . wp_json_encode( $data, JSON_UNESCAPED_SLASHES ) . ';';
+		}
+
+		// Depend on wp-i18n to ensure global `wp` exists and because anything needing this will need that too.
+		$wp_scripts->add( 'wp-jp-i18n-state', null, array( 'wp-i18n' ) );
+		$wp_scripts->add_inline_script( 'wp-jp-i18n-state', $js, 'before' );
+	}
+
+	// endregion .
+
+	// ////////////////////
+	// region Textdomain aliasing
+
+	/**
+	 * Register a textdomain alias.
+	 *
+	 * Composer packages included in plugins will likely not use the textdomain of the plugin, while
+	 * WordPress's i18n infrastructure will include the translations in the plugin's domain. This
+	 * allows for mapping the package's domain to the plugin's.
+	 *
+	 * Since multiple plugins may use the same package, we include the package's version here so
+	 * as to choose the most recent translations (which are most likely to match the package
+	 * selected by jetpack-autoloader).
+	 *
+	 * @since 1.15.0
+	 * @param string $from Domain to alias.
+	 * @param string $to Domain to alias it to.
+	 * @param string $totype What is the target of the alias: 'plugins', 'themes', or 'core'.
+	 * @param string $ver Version of the `$from` domain.
+	 * @throws InvalidArgumentException If arguments are invalid.
+	 */
+	public static function alias_textdomain( $from, $to, $totype, $ver ) {
+		if ( ! in_array( $totype, array( 'plugins', 'themes', 'core' ), true ) ) {
+			throw new InvalidArgumentException( 'Type must be "plugins", "themes", or "core"' );
+		}
+
+		if ( did_action( 'wp_default_scripts' ) ) {
+			_doing_it_wrong(
+				__METHOD__,
+				sprintf(
+					/* translators: 1: wp_default_scripts. 2: Name of the domain being aliased. */
+					esc_html__( 'Textdomain aliases should be registered before the %1$s hook. This notice was triggered by the %2$s domain.', 'jetpack' ),
+					'<code>wp_default_scripts</code>',
+					'<code>' . esc_html( $from ) . '</code>'
+				),
+				''
+			);
+		}
+
+		if ( empty( self::$domain_map[ $from ] ) ) {
+			self::init_domain_map_hooks( $from, array() === self::$domain_map );
+			self::$domain_map[ $from ] = array( $to, $totype, $ver );
+		} elseif ( Semver::compare( $ver, self::$domain_map[ $from ][2] ) > 0 ) {
+			self::$domain_map[ $from ] = array( $to, $totype, $ver );
+		}
+	}
+
+	/**
+	 * Register textdomain aliases from a mapping file.
+	 *
+	 * The mapping file is simply a PHP file that returns an array
+	 * with the following properties:
+	 *  - 'domain': String, `$to`
+	 *  - 'type': String, `$totype`
+	 *  - 'packages': Array, mapping `$from` to `$ver`.
+	 *
+	 * @since 1.15.0
+	 * @param string $file Mapping file.
+	 */
+	public static function alias_textdomains_from_file( $file ) {
+		$data = require $file;
+		foreach ( $data['packages'] as $from => $ver ) {
+			self::alias_textdomain( $from, $data['domain'], $data['type'], $ver );
+		}
+	}
+
+	/**
+	 * Register the hooks for textdomain aliasing.
+	 *
+	 * @param string $domain Domain to alias.
+	 * @param bool   $firstcall If this is the first call.
+	 */
+	private static function init_domain_map_hooks( $domain, $firstcall ) {
+		// If WordPress's plugin API is available already, use it. If not,
+		// drop data into `$wp_filter` for `WP_Hook::build_preinitialized_hooks()`.
+		if ( function_exists( 'add_filter' ) ) {
+			$add_filter = 'add_filter';
+		} else {
+			$add_filter = function ( $hook_name, $callback, $priority = 10, $accepted_args = 1 ) {
+				global $wp_filter;
+				// phpcs:ignore WordPress.WP.GlobalVariablesOverride.Prohibited
+				$wp_filter[ $hook_name ][ $priority ][] = array(
+					'accepted_args' => $accepted_args,
+					'function'      => $callback,
+				);
+			};
+		}
+
+		$add_filter( "gettext_{$domain}", array( self::class, 'filter_gettext' ), 10, 3 );
+		$add_filter( "ngettext_{$domain}", array( self::class, 'filter_ngettext' ), 10, 5 );
+		$add_filter( "gettext_with_context_{$domain}", array( self::class, 'filter_gettext_with_context' ), 10, 4 );
+		$add_filter( "ngettext_with_context_{$domain}", array( self::class, 'filter_ngettext_with_context' ), 10, 6 );
+		if ( $firstcall ) {
+			$add_filter( 'load_script_translation_file', array( self::class, 'filter_load_script_translation_file' ), 10, 3 );
+		}
+	}
+
+	/**
+	 * Filter for `gettext`.
+	 *
+	 * @since 1.15.0
+	 * @param string $translation Translated text.
+	 * @param string $text Text to translate.
+	 * @param string $domain Text domain.
+	 * @return string Translated text.
+	 */
+	public static function filter_gettext( $translation, $text, $domain ) {
+		if ( $translation === $text ) {
+			// phpcs:ignore WordPress.WP.I18n
+			$newtext = __( $text, self::$domain_map[ $domain ][0] );
+			if ( $newtext !== $text ) {
+				return $newtext;
+			}
+		}
+		return $translation;
+	}
+
+	/**
+	 * Filter for `ngettext`.
+	 *
+	 * @since 1.15.0
+	 * @param string $translation Translated text.
+	 * @param string $single The text to be used if the number is singular.
+	 * @param string $plural The text to be used if the number is plural.
+	 * @param string $number The number to compare against to use either the singular or plural form.
+	 * @param string $domain Text domain.
+	 * @return string Translated text.
+	 */
+	public static function filter_ngettext( $translation, $single, $plural, $number, $domain ) {
+		if ( $translation === $single || $translation === $plural ) {
+			// phpcs:ignore WordPress.WP.I18n
+			$translation = _n( $single, $plural, $number, self::$domain_map[ $domain ][0] );
+		}
+		return $translation;
+	}
+
+	/**
+	 * Filter for `gettext_with_context`.
+	 *
+	 * @since 1.15.0
+	 * @param string $translation Translated text.
+	 * @param string $text Text to translate.
+	 * @param string $context Context information for the translators.
+	 * @param string $domain Text domain.
+	 * @return string Translated text.
+	 */
+	public static function filter_gettext_with_context( $translation, $text, $context, $domain ) {
+		if ( $translation === $text ) {
+			// phpcs:ignore WordPress.WP.I18n
+			$translation = _x( $text, $context, self::$domain_map[ $domain ][0] );
+		}
+		return $translation;
+	}
+
+	/**
+	 * Filter for `ngettext_with_context`.
+	 *
+	 * @since 1.15.0
+	 * @param string $translation Translated text.
+	 * @param string $single The text to be used if the number is singular.
+	 * @param string $plural The text to be used if the number is plural.
+	 * @param string $number The number to compare against to use either the singular or plural form.
+	 * @param string $context Context information for the translators.
+	 * @param string $domain Text domain.
+	 * @return string Translated text.
+	 */
+	public static function filter_ngettext_with_context( $translation, $single, $plural, $number, $context, $domain ) {
+		if ( $translation === $single || $translation === $plural ) {
+			// phpcs:ignore WordPress.WP.I18n
+			$translation = _nx( $single, $plural, $number, $context, self::$domain_map[ $domain ][0] );
+		}
+		return $translation;
+	}
+
+	/**
+	 * Filter for `load_script_translation_file`.
+	 *
+	 * @since 1.15.0
+	 * @param string|false $file Path to the translation file to load. False if there isn't one.
+	 * @param string       $handle Name of the script to register a translation domain to.
+	 * @param string       $domain The text domain.
+	 */
+	public static function filter_load_script_translation_file( $file, $handle, $domain ) {
+		if ( false !== $file && isset( self::$domain_map[ $domain ] ) && ! is_readable( $file ) ) {
+			// Determine the part of the filename after the domain.
+			$suffix = basename( $file );
+			$l      = strlen( $domain );
+			if ( substr( $suffix, 0, $l ) !== $domain || '-' !== $suffix[ $l ] ) {
+				return $file;
+			}
+			$suffix   = substr( $suffix, $l );
+			$lang_dir = Jetpack_Constants::get_constant( 'WP_LANG_DIR' );
+
+			// Look for replacement files.
+			list( $newdomain, $type ) = self::$domain_map[ $domain ];
+			$newfile                  = $lang_dir . ( 'core' === $type ? '/' : "/{$type}/" ) . $newdomain . $suffix;
+			if ( is_readable( $newfile ) ) {
+				return $newfile;
+			}
+		}
+		return $file;
+	}
+
+	// endregion .
+
 }
+
+// Enable section folding in vim:
+// vim: foldmarker=//\ region,//\ endregion foldmethod=marker
+// .