+++ projects/plugins/jetpack/class.jetpack-plan.php
@@ -103,7 +103,6 @@ class Jetpack_Plan {
 				'ecommerce-bundle',
 				'ecommerce-bundle-monthly',
 				'ecommerce-bundle-2y',
-				'vip',
 			),
 			'supports' => array(),
 		),
@@ -112,6 +111,7 @@ class Jetpack_Plan {
 			'plans'    => array(
 				'jetpack_complete',
 				'jetpack_complete_monthly',
+				'vip',
 			),
 			'supports' => array(),
 		),