+++ projects/plugins/jetpack/extensions/blocks/premium-content/login-button/color-edit.js
@@ -54,7 +54,7 @@ function ColorPanel( { settings, clientId, enableContrastChecking = true } ) {
 
 	const title = isWebPlatform
 		? __( 'Color settings', 'jetpack' )
-		: __( 'Color Settings', 'jetpack' );
+		: __( 'Color Settings', 'jetpack', /* dummy arg to avoid bad minification */ 0 );
 
 	useEffect( () => {
 		if ( isWebPlatform && ! enableContrastChecking ) {