+++ projects/js-packages/idc/CHANGELOG.md
@@ -2,6 +2,28 @@
 
 ### This is a list detailing changes for the Jetpack RNA IDC package releases.
 
+## 0.7.0 - 2021-12-14
+### Added
+- Add the non-admin IDC screen.
+
+## 0.6.0 - 2021-12-07
+### Added
+- Add event tracking.
+- IDC screen content customization.
+
+### Changed
+- Extract visual components and create Storybook stories.
+- Updated package dependencies.
+
+## 0.5.1 - 2021-11-30
+### Changed
+- Updated package dependencies.
+
+## 0.5.0 - 2021-11-23
+### Changed
+- Add "Staying in safe mode" spinner, rename CSS classes according to the CSS guidelines.
+- Updated package dependencies
+
 ## 0.4.4 - 2021-11-17
 ### Changed
 - Updated package dependencies.