+++ projects/packages/analyzer/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.2] - 2021-12-07
+### Added
+- Handle more code tokens
+
+### Changed
+- Changes to Analyzer serialization
+- Updated package dependencies.
+
 ## [1.6.1] - 2021-08-26
 ### Added
 - Composer alias for dev-master, to improve dependencies.
@@ -60,6 +68,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Jetpack code analyzer
 
+[1.6.2]: https://github.com/Automattic/jetpack-analyzer/compare/v1.6.1...v1.6.2
 [1.6.1]: https://github.com/Automattic/jetpack-analyzer/compare/v1.6.0...v1.6.1
 [1.6.0]: https://github.com/Automattic/jetpack-analyzer/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/Automattic/jetpack-analyzer/compare/v1.4.0...v1.5.0